+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 2020-05-10
 - IR.CreateDevice - added
 - IR.mockAppStart should be called to call IR.EVENT_START and connect all added devices
+- IR.mockResetIr - to reset IR
 - added jest tests (timers, variables, devices)
 
 ## 2020-01-15